Work-in-progress refactor.
Major change is that extension lifecycle actions are encapsulated in
a new Extension model provided by a new Extension service.

Note: This code has not actually been executed. It is better than
pseudocode, but it may not run without edits, and the controller
will certainly need to updated to incorporate the new model and
service.