Cleaned up version tracking in the Extension object.
After some discussion, we determined version doesn't need to be tracked
in stateOnPage. The version on the page and the version in the database
will always match. The purpose of the stateOnPage object is to track
differences between the extension in the database and the extension
as represented on the screen.

... and future commits might factor out stateOnPage altogether.